ectc-project/gateway/requirements.txt
# Core scientific computing
numpy==1.21.0
scipy==1.7.0
scikit-learn==1.0.0

# Machine learning
torch==1.9.0
tensorflow-cpu==2.6.0

# JIT compilation
numba==0.54.0

# Cryptography
pycryptodome==3.15.0

# Serial communication
pyserial==3.5

# MQTT (optional)
paho-mqtt==1.6.1

# Database (optional)
influxdb-client==1.30.0

# Plotting
matplotlib==3.4.3
seaborn==0.11.2

# Utilities
tqdm==4.62.3
pyyaml==6.0
psutil==5.8.0
